feat(sidebar): replace Todo section with Plan in left sidebar
Show plan_steps progress and goal status instead of todo_write items.
Aligns with Plan mode workflow where the sidebar tracks design progress
rather than model-managed task lists.
